% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_text_geom}
\alias{add_text_geom}
\title{Add Data Labels to Plot}
\usage{
add_text_geom(
  label_type = "count",
  accuracy = 0.1,
  color = "white",
  vjust = 1.6,
  fontface = "bold"
)
}
\arguments{
\item{fontface}{}
}
\description{
Add Data Labels to Plot
}
