% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{badged_n_days}
\alias{badged_n_days}
\title{Badged Days}
\usage{
badged_n_days(trainee_data, days_threshold = 90)
}
\arguments{
\item{trainee_data}{a data frame}

\item{days_threshold}{the number of days past training (default = 90)}
}
\value{
a single number
}
\description{
Badged Days
}
