% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_avg_time_to_checkout}
\alias{calculate_avg_time_to_checkout}
\title{Calculate average time to checkout}
\usage{
calculate_avg_time_to_checkout(trainee_progress, days_threshold = 90)
}
\arguments{
\item{trainee_progress}{data frame}

\item{days_threshold}{number of days since training, default = 90}
}
\value{
average number of days
}
\description{
Calculate average time to checkout
}
\examples{
trainee_progress <- fetch_redash(388)
(calculate_avg_time_to_checkout(trainee_progress))
}
