% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_checkout_not_finished}
\alias{calculate_checkout_not_finished}
\title{Title}
\usage{
calculate_checkout_not_finished(trainee_progress, days_threshold = 90)
}
\arguments{
\item{days_threshold}{}
}
\description{
Title
}
