% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{calculate_nps}
\alias{calculate_nps}
\title{Calculate Net Promoter Score}
\usage{
calculate_nps(dat, col)
}
\arguments{
\item{col}{}
}
\description{
#'How to calculate net promoter score: https://monkeylear.com/blog/nps-analysis/
}
