% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_reengagement_rate}
\alias{calculate_reengagement_rate}
\title{Calculate Re-engagement rate}
\usage{
calculate_reengagement_rate(trainee_data, days_threshold = 180)
}
\arguments{
\item{trainee_data}{a data frame}

\item{days_threshold}{the number of days you would like to review, default = 180 days (6 months)}
}
\value{
percentage
}
\description{
Calculate Re-engagement rate
}
