% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_wksurvey_item_mean}
\alias{calculate_wksurvey_item_mean}
\title{Calculate Workshop Survey Item Mean}
\usage{
calculate_wksurvey_item_mean(data, item)
}
\arguments{
\item{data}{dataframe}

\item{item}{column name}
}
\description{
Calculate Workshop Survey Item Mean
}
