% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{count_trainings_yr}
\alias{count_trainings_yr}
\title{Count Trainings by Year}
\usage{
count_trainings_yr(dat)
}
\arguments{
\item{dat}{a data frame}
}
\value{
a data frame
}
\description{
Count Trainings by Year
}
