% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{custom_scale_x_discrete}
\alias{custom_scale_x_discrete}
\title{Format and Text Wrap X Axis Labels}
\usage{
custom_scale_x_discrete(labels, wrap_width = 15)
}
\arguments{
\item{wrap_width}{}
}
\description{
Format and Text Wrap X Axis Labels
}
