% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_airtable}
\alias{fetch_airtable}
\title{Fetch Airtable Data}
\usage{
fetch_airtable(id, db_name)
}
\arguments{
\item{id}{the base id. Can be found in Airtable API documentation}

\item{db_name}{the Database name. Can be found in Airtable API documentation.}
}
\value{
data frame
}
\description{
Fetch Airtable Data
}
