% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{generate_plot}
\alias{generate_plot}
\title{Generate Plots}
\usage{
generate_plot(dat, x, y, fill = NULL)
}
\arguments{
\item{dat}{data frame}

\item{x}{variable on x axis}

\item{y}{variable on y axis}

\item{fill}{variable to be used at fill}
}
\value{
plot
}
\description{
Generate Plots
}
