% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_env_var}
\alias{get_env_var}
\title{Retrieve Environment Variable Value}
\usage{
get_env_var(var)
}
\arguments{
\item{var}{Character; The name of the environment variable to retrieve.}
}
\value{
The value of the specified environment variable as a character string.
Throws an error if the environment variable is not set.
}
\description{
This function fetches the value of a specified environment variable.
If the environment variable is not set, the function will stop execution and return an error message.
}
\examples{
\dontrun{
  # Set an example environment variable
  Sys.setenv(EXAMPLE_VAR = "example_value")

  # Fetch the value of the environment variable
  val <- get_env_var("EXAMPLE_VAR")
  print(val)  # Output should be "example_value"
}
}
