% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{group_by_summarise}
\alias{group_by_summarise}
\title{Group Data by a Specified Column and Summarize Counts}
\usage{
group_by_summarise(dat, col, convert_to_percent = FALSE)
}
\arguments{
\item{dat}{A data frame to be analyzed.}

\item{col}{A character string specifying the column name by which to group the data.}

\item{convert_to_percent}{A logical value; if TRUE, convert counts to percentages.
Default is FALSE.}
}
\value{
A data frame with each group (from the specified column) and its count (and optionally, percentage).
}
\description{
This function groups a data frame by the values in a specified column, then calculates
the count of records in each group. Optionally, it can convert these counts into percentages.
}
\examples{
# Example usage with a sample data frame
set.seed(1)  # for reproducibility
sample_data <- data.frame(category = sample(c("A", "B", "C"), 120, replace = TRUE, prob = c(0.3, 0.5, 0.2)))
summary_table <- group_by_summarise(sample_data, "category", convert_to_percent = TRUE)
print(summary_table)
}
