% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{map_availability}
\alias{map_availability}
\title{Map Trainer Availability}
\usage{
map_availability(dat)
}
\arguments{
\item{dat}{a data frame}
}
\value{
data frame with availability column simplified
}
\description{
Map Trainer Availability
}
