% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{map_scheduling_group}
\alias{map_scheduling_group}
\title{Create Scheduling Groups}
\usage{
map_scheduling_group(df)
}
\arguments{
\item{df}{Dataframe; The data frame containing timezone information. Comes from Trainer Quarterly Availability Form.}
}
\value{
Dataframe; The data frame with an additional 'sched_group'
column.
Throws an error if no matching column or multiple matching columns are found.
}
\description{
Create Scheduling Groups
}
\examples{
\dontrun{
  # Assume 'data_with_timezones' is a data frame with an appropriate timezone column
  grouped_timezones <- map_timezones(data_with_timezones)
  print(grouped_timezones)
}

}
