% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{preprocess_trainee_data}
\alias{preprocess_trainee_data}
\title{Preprocess Trainee Data}
\usage{
preprocess_trainee_data(trainee_data)
}
\arguments{
\item{trainee_data}{data.frame
A data frame containing the raw trainee data. Expected to have columns named "training" and "instr_badge",
which represent the training and instruction badging dates, respectively.}
}
\value{
data.frame
A preprocessed data frame with added columns:
\itemize{
\item \code{days_from_training}: Number of days from training date to today's date.
\item \code{days_to_badge}: Number of days from training date to badging date.
}
}
\description{
This function preprocesses the trainee data by converting columns to dates and calculating the number of days
from the training date to the current date (\code{days_from_training}) and from the training date to the badging date (\code{days_to_badge}).
}
\examples{
\dontrun{
  raw_trainee_data <- data.frame(
    training = c("2023-01-01", "2023-01-10"),
    instr_badge = c("2023-02-01", NA)
  )
  preprocessed_data <- preprocess_trainee_data(raw_trainee_data)
  print(preprocessed_data)
}
}
