% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{rename_and_group_availability}
\alias{rename_and_group_availability}
\title{Rename and Group Trainer Availability Data}
\usage{
rename_and_group_availability(dat)
}
\arguments{
\item{dat}{}
}
\description{
Rename and Group Trainer Availability Data
}
\examples{
sample_data <- googlesheets4::read_sheet("1epfbcCrR2WsaaDo_rNNvX8GrSJK0Px2m9ph6Oj6sEhI")

result <- rename_and_group_availability(sample_data)

str(result)
}
