% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plot}
\alias{save_plot}
\title{Save and Export ggplot2 Plots to a Specified Directory}
\usage{
save_plot(dir, plot)
}
\arguments{
\item{dir}{Character; The name of the subdirectory under 'programs' where the plot will be saved.}

\item{plot}{ggplot; The ggplot2 plot object that you want to save.}
}
\value{
Saves the plot as a PNG file and returns the full path where the file is saved.
}
\description{
This function saves a ggplot2 plot as a PNG file in a specified directory.
The plot will be saved under the 'programs/\if{html}{\out{<dir>}}/output/' folder with the provided plot name.
}
\examples{
\dontrun{
  # Assume 'some_plot' is a ggplot2 plot object
  saved_path <- save_plot(dir = "exampleDir", plot = some_plot)
  print(saved_path)
}
}
