% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{set_time_frame}
\alias{set_time_frame}
\title{Set Time Frame}
\usage{
set_time_frame(dat, col, start, end)
}
\arguments{
\item{dat}{a dataframe}

\item{col}{a column in the dataframe}

\item{start}{the date you want to start the time frame}

\item{end}{the date you want to end the time frame}
}
\value{
the dataframe filtered by the time range
}
\description{
Set Time Frame
}
