% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{summarise_data_by_tz}
\alias{summarise_data_by_tz}
\title{Summarise Trainer Availability Data by Timezone Group}
\usage{
summarise_data_by_tz(dat)
}
\arguments{
\item{dat}{data frame}
}
\value{
a data frame
}
\description{
Summarise Trainer Availability Data by Timezone Group
}
