% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_checkout_not_finished}
\alias{calculate_checkout_not_finished}
\title{Calculate checkout not finished}
\usage{
calculate_checkout_not_finished(trainee_progress, days_threshold = 90)
}
\arguments{
\item{trainee_progress}{a data frame}

\item{days_threshold}{the number of days past training (default = 90)}
}
\value{
number of people who have started but not finished checkout
}
\description{
Calculate checkout not finished
}
