% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_checkout_not_finished}
\alias{calculate_checkout_not_finished}
\title{Calculate Number of Trainees Who Started but Haven't Finished Checkout}
\usage{
calculate_checkout_not_finished(trainee_progress, days_threshold = 90)
}
\arguments{
\item{trainee_progress}{data.frame
A data frame containing progress indicators for each trainee.}

\item{days_threshold}{integer
The number of days since training to consider. Default is 90 days.}
}
\value{
integer
Returns the number of trainees who have started but not finished the checkout process.
}
\description{
This function calculates the number of trainees who started but haven't completed the checkout process within a specified number of days since their training.
}
\examples{
\dontrun{
  raw_data <- data.frame(...)
  n_not_finished_checkout <- calculate_checkout_not_finished(raw_data, 90)
}
}
