% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_checkout_rate}
\alias{calculate_checkout_rate}
\title{Calculate checkout rate}
\usage{
calculate_checkout_rate(trainee_data, days_threshold = 90)
}
\arguments{
\item{days_threshold}{}
}
\value{
percentage
}
\description{
Calculate checkout rate
}
