% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_checkout_rate}
\alias{calculate_checkout_rate}
\title{Calculate checkout rate}
\usage{
calculate_checkout_rate(trainee_data, days_threshold = 90)
}
\arguments{
\item{trainee_data}{a data frame}

\item{days_threshold}{the number of days you would like to review, default = 90 days}
}
\value{
percentage
}
\description{
Calculate checkout rate
}
