% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_checkout_started}
\alias{calculate_checkout_started}
\title{Calculate Checkout Started}
\usage{
calculate_checkout_started(trainee_progress, days_threshold = 90)
}
\arguments{
\item{days_threshold}{}
}
\value{
number of people who have started checkout
}
\description{
Calculate Checkout Started
}
