% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_checkout_started}
\alias{calculate_checkout_started}
\title{Calculate Number of Trainees Who Have Started Checkout}
\usage{
calculate_checkout_started(trainee_progress, days_threshold = 90)
}
\arguments{
\item{trainee_progress}{data.frame
A data frame containing checkout steps for each trainee, like 'get_involved', 'teaching_demo', and 'welcome'.}

\item{days_threshold}{integer
The number of days since training to consider. Default is 90 days.}
}
\value{
integer
Returns the number of trainees who have started the checkout process.
}
\description{
This function calculates the number of trainees who have started the checkout process within a specified number of days since their training.
}
\examples{
\dontrun{
  raw_data <- data.frame(...)
  n_started_checkout <- calculate_checkout_started(raw_data, 90)
}
}
