% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_dropout_rate}
\alias{calculate_dropout_rate}
\title{Instructor Checkout Dropout Rate}
\usage{
calculate_dropout_rate(trainee_progress, days_threshold = 90)
}
\arguments{
\item{days_threshold}{}
}
\value{
percentage
}
\description{
Instructor Checkout Dropout Rate
}
