% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{calculate_rate}
\alias{calculate_rate}
\title{Calculate a rate}
\usage{
calculate_rate(xsmaller, ylarger)
}
\arguments{
\item{xsmaller}{the smaller number}

\item{ylarger}{the larger number}
}
\value{
a percentage
}
\description{
Calculate a rate
}
\examples{
calculate_rate(50, 100)
}
