% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{calculate_wksurvey_item_mean}
\alias{calculate_wksurvey_item_mean}
\title{Calculate Workshop Survey Item Mean}
\usage{
calculate_wksurvey_item_mean(data, item)
}
\arguments{
\item{data}{dataframe}

\item{item}{column name}
}
\description{
Calculate Workshop Survey Item Mean
}
\examples{
 mock_data <- data.frame(submitted_at = as.Date(c("2023-08-15", "2023-08-13")), instructors_clear_answers = c(4, 5))
result <- calculate_wksurvey_item_mean(mock_data, "instructors_clear_answers")
print(result)


}
