% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{clean_select_all_responses}
\alias{clean_select_all_responses}
\title{Clean Select All Survey Response Data}
\usage{
clean_select_all_responses(dat, col)
}
\arguments{
\item{col}{}
}
\description{
Clean Select All Survey Response Data
}
