% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{count_trainings}
\alias{count_trainings}
\title{Number of Trainings Trainers Have Taught}
\usage{
count_trainings(dat)
}
\arguments{
\item{dat}{a data frame}
}
\value{
a data frame
}
\description{
Number of Trainings Trainers Have Taught
}
