% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{extract_year_quarter}
\alias{extract_year_quarter}
\title{Extract Quarter and Year from Filename}
\usage{
extract_year_quarter(df)
}
\arguments{
\item{filename}{}
}
\description{
Extract Quarter and Year from Filename
}
\examples{
Q1_2034 <- data.frame()
result <- extract_year_quarter(Q1_2034)
print(result)
}
