% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_redash}
\alias{fetch_redash}
\title{Fetch Redash Data}
\usage{
fetch_redash(query_id)
}
\arguments{
\item{query_id}{Can be found in redash.}
}
\value{
the data frame from the redash query supplied
}
\description{
Fetch Redash Data
}
\examples{

test <- fetch_redash(17)

}
