% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{gather_and_summarise}
\alias{gather_and_summarise}
\title{Gather Questions into One Table to Graph Together}
\usage{
gather_and_summarise(dat, questions, levels, labels)
}
\arguments{
\item{labels}{}
}
\description{
Gather Questions into One Table to Graph Together
}
