% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{generate_plot}
\alias{generate_plot}
\title{This function generates a bar plot using \code{ggplot2} and a custom Carpentries theme.
It allows for specifying variables for the x-axis, y-axis, and fill color. The plot is faceted by year and quarter.
Text labels indicating the y-values are also included above the bars.}
\usage{
generate_plot(dat, x, y, fill = NULL)
}
\arguments{
\item{dat}{Dataframe; The data frame containing the data to be plotted.}

\item{x}{Character; The name of the variable to be plotted on the x-axis.Character; The name of the variable to be plotted on the x-axis.}

\item{y}{Character; The name of the variable to be plotted on the y-axis.}

\item{fill}{Character; Optional. The name of the variable to be used for fill colors.}
}
\value{
A ggplot object representing the generated plot.
}
\description{
This function generates a bar plot using \code{ggplot2} and a custom Carpentries theme.
It allows for specifying variables for the x-axis, y-axis, and fill color. The plot is faceted by year and quarter.
Text labels indicating the y-values are also included above the bars.
}
\examples{
\dontrun{
  # Assume 'data_frame' is a data frame with appropriate columns
  example_plot <-
  generate_plot(dat = data_frame, x = "variable_x",
  y = "variable_y", fill = "variable_fill")
  print(example_plot)
}

}
