% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{group_by_summarise}
\alias{group_by_summarise}
\title{Group By and Summarise}
\usage{
group_by_summarise(dat, col, convert_to_percent = FALSE)
}
\arguments{
\item{convert_to_percent}{}
}
\description{
Group By and Summarise
}
