% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{map_availability}
\alias{map_availability}
\title{Map Trainer Availability}
\usage{
map_availability(dat)
}
\arguments{
\item{dat}{}
}
\value{
data frame with availability column simplified
}
\description{
Map Trainer Availability
}
\examples{
sample_data <- googlesheets4::read_sheet("1epfbcCrR2WsaaDo_rNNvX8GrSJK0Px2m9ph6Oj6sEhI")

result <- map_availability(sample_data)
str(result)
}
