% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{map_timezones}
\alias{map_timezones}
\title{Map Timezones}
\usage{
map_timezones(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
data frame with timezones simplified
}
\description{
Map Timezones
}
