% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{map_timezones}
\alias{map_timezones}
\title{Map Timezones}
\usage{
map_timezones(dat)
}
\arguments{
\item{dat}{a dataframe}
}
\value{
dataframe with timezones simplified
}
\description{
Map Timezones
}
\examples{
sample_data <- googlesheets4::read_sheet("1epfbcCrR2WsaaDo_rNNvX8GrSJK0Px2m9ph6Oj6sEhI")

result <-
map_timezones(sample_data)
print(result)
}
