% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plot}
\alias{save_plot}
\title{Save and Export Plots}
\usage{
save_plot(dir, plot)
}
\arguments{
\item{dir}{a file directory to save the plot}

\item{plot}{the name of the plot}
}
\value{
a file
}
\description{
Save and Export Plots
}
