% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{set_time_frame}
\alias{set_time_frame}
\title{Filter Data Frame by start and end date}
\usage{
set_time_frame(dat, col, start, end)
}
\arguments{
\item{dat}{a dataframe}

\item{col}{a column in the dataframe}

\item{start}{the date you want to start the time frame}

\item{end}{the date you want to end the time frame}
}
\value{
the dataframe filtered by the time range
}
\description{
Filter Data Frame by start and end date
}
\examples{
trainee_progress <- fetch_redash(388)
result <- set_time_frame(dat = trainee_progress, col = "training", start = "2021-01-01", end = "2021-12-31")
head(result)
}
