% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{summarise_data_by_qtr}
\alias{summarise_data_by_qtr}
\title{Summarise Trainer Availability Data by Quarter}
\usage{
summarise_data_by_qtr(dat)
}
\arguments{
\item{dat}{}
}
\description{
Summarise Trainer Availability Data by Quarter
}
\examples{
q1_2034 <- as.data.frame(googlesheets4::read_sheet("1epfbcCrR2WsaaDo_rNNvX8GrSJK0Px2m9ph6Oj6sEhI"))

result <- summarise_data_by_qtr(q1_2034)

print(result)

To do: Make this work with real quarter and year instead of 2034 01
}
