% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructor_training.R
\name{summarise_data_by_tz}
\alias{summarise_data_by_tz}
\title{Summarise Trainer Availability Data by Timezone Group}
\usage{
summarise_data_by_tz(dat)
}
\arguments{
\item{dat}{}
}
\description{
Summarise Trainer Availability Data by Timezone Group
}
\examples{
q1_2034 <- as.data.frame(googlesheets4::read_sheet("1epfbcCrR2WsaaDo_rNNvX8GrSJK0Px2m9ph6Oj6sEhI"))

result <- summarise_data_by_tz(q1_2034)

print(result)
}
